Cleanup tests for issue 1724
Remove old XFAIL test for 1724 and move the new one to test_complex.py
Move test for 2330 as well.

NB: issues 1724 and 2330 have been solved
by 8efe52a6a4acf8307e4e0dd176ccce720604cce3
