+++ sympy/core/tests/test_arit.py
@@ -164,14 +164,6 @@ def test_pow3():
     assert sqrt(2)**3 == 2 * sqrt(2)
     assert sqrt(2)**3 == sqrt(8)
 
-def test_pow_issue_1724():
-    e = ((-1)**(S(1)/3))
-    assert e.conjugate().n() == e.n().conjugate()
-    e = S('-2/3 - (-29/54 + sqrt(93)/18)**(1/3) - 1/(9*(-29/54 + sqrt(93)/18)**(1/3))')
-    assert e.conjugate().n() == e.n().conjugate()
-    e = 2**I
-    assert e.conjugate().n() == e.n().conjugate()
-
 def test_expand():
     p = Rational(5)
     e = (a+b)*c
